+++ client/src/data-sources/plugins/ApplicationInsights/Query.ts
@@ -2,7 +2,7 @@ import * as request from 'xhr-request';
 import { DataSourcePlugin, IOptions } from '../DataSourcePlugin';
 import { appInsightsUri } from './common';
 import ApplicationInsightsConnection from '../../connections/application-insights';
-import { DataSourceConnector } from '../../DataSourceConnector';
+import { DataSourceConnector, IDataSource } from '../../DataSourceConnector';
 
 let connectionType = new ApplicationInsightsConnection();
 
@@ -164,6 +164,22 @@ export default class ApplicationInsightsQuery extends DataSourcePlugin<IQueryPar
     }
   }
 
+  getElementQuery(dataSource: IDataSource, dependencies: IDict<any>, aQuery: string, queryFilters: any): string {
+    let timespan = '30d';
+    const table = dataSource['config'].params.table || null;
+    if (dependencies['timespan'] && dependencies['timespan']['queryTimespan']) {
+      timespan = dependencies['timespan']['queryTimespan'];
+      timespan = this.convertApplicationInsightsTimespan(timespan);
+    } else if (dependencies['queryTimespan']) {
+      // handle dialog params
+      timespan = dependencies['queryTimespan'];
+      timespan = this.convertApplicationInsightsTimespan(timespan);
+    }
+    const filter = this.formatApplicationInsightsFilterString(queryFilters, dependencies);
+    const query = this.formatApplicationInsightsQueryString(aQuery, timespan, filter, table);
+    return query;
+  }
+
   private mapAllTables(results: IQueryResults, mappings: Array<IDictionary>): any[][] {
 
     if (!results || !results.Tables || !results.Tables.length) {
@@ -259,4 +275,51 @@ export default class ApplicationInsightsQuery extends DataSourcePlugin<IQueryPar
       throw new Error('{ table, queries } should be of types { "string", { query1: {...}, query2: {...} }  }.');
     }
   }
+
+  // element export formatting functions
+  private formatApplicationInsightsQueryString(query: string, timespan: string, filter: string, table?: string) {
+    let str = query.replace(/(\|)((\s??\n\s??)(.+?))/gm, '\n| $4'); // move '|' to start of each line
+    str = str.replace(/(\s?\|\s?)([^\|])/gim, '\n| $2'); // newline on '|'
+    str = str.replace(/(\sand\s)/gm, '\n $1'); // newline on 'and'
+    str = str.replace(/([^\(\'\"])(,\s*)(?=(\w+[^\)\'\"]\w+))/gim, '$1,\n  '); // newline on ','
+    const timespanQuery = '| where timestamp > ago(' + timespan + ') \n';
+    // Checks for '|' at start
+    const matches = str.match(/^(\s*\||\s*\w+\s*\|)/ig); 
+    const start = (!matches || matches.length !== 1) ? '| ' : '';
+    if (table) {
+      str = table + ' \n' + timespanQuery + filter + start + str;
+    } else {
+      // Insert timespan and filter after table name
+      str = str.replace(/^\s*(\w+)\s*/gi, '$1 \n' + timespanQuery + filter + start); 
+    }
+    return str;
+  }
+
+  private formatApplicationInsightsFilterString(filters: IStringDictionary[], dependencies: IDict<any>) {
+    let str = '';
+    if (!filters) {
+      return str;
+    }
+    // Apply selected filters to connected query
+    filters.forEach((filter) => {
+      const { dependency, queryProperty } = filter;
+      const selectedFilters: string[] = dependencies[dependency] || [];
+      if (selectedFilters.length > 0) {
+        const f = '| where ' + selectedFilters.map((value) => `${queryProperty}=="${value}"`).join(' or ');
+        str = `${str}${f} \n`;
+      }
+    });
+    return str;
+  }
+
+  private convertApplicationInsightsTimespan(timespan: string) {
+    let str = timespan;
+    if (timespan.substr(0, 2) === 'PT') {
+      str = str.substring(2).toLowerCase();
+    } else if (timespan.substr(0, 1) === 'P') {
+      str = str.substring(1).toLowerCase();
+    }
+    return str;
+  }
+
 }
